Rename dashboard -> analytics; remove cruft
