+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-testpaths = stats/tests statsrunner tests
+testpaths = stats/tests statsrunner
 norecursedirs = *__pycache__* *.pytest_cache*
 console_output_style = count
