+++ stats/tests/test_comprehensiveness.py
@@ -4,7 +4,7 @@ from lxml import etree
 import datetime
 import pytest
 
-from stats.dashboard import ActivityStats
+from stats.analytics import ActivityStats
 
 
 class MockActivityStats(ActivityStats):
